@@ -20,22 +20,22 @@
     ],
     "homepage": "https://github.com/Riverline/worker-bundle",
     "require":{
-        "php":                          ">=5.6",
+        "php":                          ">=7.1.3",
         "ext-zlib":                     "*",
-        "symfony/console":              "^2.0|^3.0",
-        "symfony/dependency-injection": "^2.0|^3.0",
-        "symfony/config":               "^2.0|^3.0",
-        "symfony/http-kernel":          "^2.0|^3.0"
+        "symfony/console":              "^4.0",
+        "symfony/dependency-injection": "^4.0",
+        "symfony/config":               "^4.0",
+        "symfony/http-kernel":          "^4.0"
     },
     "suggest": {
         "amazonwebservices/aws-sdk-for-php": "^1.0",
-        "aws/aws-sdk-php":                   "^2.0",
+        "aws/aws-sdk-php":                   "^3.0",
         "predis/predis":                     "^1.0",
         "stomp-php/stomp-php":               "^4.0"
     },
     "require-dev": {
         "phpunit/phpunit":     "^4.8",
-        "aws/aws-sdk-php":     "^2.8",
+        "aws/aws-sdk-php":     "^3.0",
         "predis/predis":       "^1.1",
         "stomp-php/stomp-php": "^4.3"
     },
