@@ -2,68 +2,74 @@
 
 namespace Riverline\WorkerBundle\Provider;
 
+use Riverline\WorkerBundle\Queue\Queue;
+
 /**
  * Class AwsSQSv2Test
  * @package Riverline\WorkerBundle\Provider
+ * @deprecated
  */
 class AwsSQSv2Test extends \PHPUnit_Framework_TestCase
 {
 
     /**
-     * @var \Riverline\WorkerBundle\Provider\AwsSQSv2
+     * @var AwsSQSv2
      */
     private $provider;
 
-    public function setUp()
+    public function setUp(): void
     {
-        $this->provider = new AwsSQSv2(array(
-            'key'      => "key",
-            'secret'   => "secret",
-            'region'   => "us-east-1",
-            'version'  => "latest",
-            'endpoint' => 'http://aws:4576',
-        ));
+        self::markTestSkipped("SDK v2 is used");
+        $this->provider = new AwsSQSv2(
+            [
+                'key' => "key",
+                'secret' => "secret",
+                'region' => "us-east-1",
+                'version' => "latest",
+                'endpoint' => 'http://aws:4566',
+            ]
+        );
     }
 
-    public function testCreateQueue()
+    public function testCreateQueue(): void
     {
-        $newQueue = $this->provider->createQueue("RiverlineWorkerBundleTest_create_v2", array('VisibilityTimeout' => '15'));
+        $newQueue = $this->provider->createQueue("RiverlineWorkerBundleTest_create_v2", ['VisibilityTimeout' => '15']);
 
-        $this->assertTrue($newQueue instanceof \Riverline\WorkerBundle\Queue\Queue);
+        self::assertInstanceOf(Queue::class, $newQueue);
     }
 
-    public function testPutArray()
+    public function testPutArray(): void
     {
-        $this->provider->put("RiverlineWorkerBundleTest_create_v2", array('name' => 'Romain'));
+        $this->provider->put("RiverlineWorkerBundleTest_create_v2", ['name' => 'Romain']);
     }
 
-    public function testCount()
+    public function testCount(): void
     {
         $count = $this->provider->count("RiverlineWorkerBundleTest_create_v2");
 
-        $this->assertEquals(1, $count);
+        self::assertEquals(1, $count);
     }
 
-    public function testGetArray()
+    public function testGetArray(): void
     {
         $workload = $this->provider->get("RiverlineWorkerBundleTest_create_v2");
 
-        $this->assertSame(array('name' => 'Romain'), $workload);
+        self::assertSame(['name' => 'Romain'], $workload);
     }
 
-    public function testTimeout()
+    public function testTimeout(): void
     {
         $tic = time();
 
         $this->provider->get("RiverlineWorkerBundleTest_create_v2", 3);
 
-        $this->assertGreaterThanOrEqual(3, time() - $tic);
+        self::assertGreaterThanOrEqual(3, time() - $tic);
     }
 
-    public function testMultiPut()
+    public function testMultiPut(): void
     {
-        $workloads = array();
-        for($i = 0 ; $i < 10 ; $i++) {
+        $workloads = [];
+        for ($i = 0; $i < 10; $i++) {
             $workloads[] = "workload$i";
         }
 
@@ -73,75 +79,83 @@ class AwsSQSv2Test extends \PHPUnit_Framework_TestCase
 
         $count = $this->provider->count("RiverlineWorkerBundleTest_create_v2");
 
-        $this->assertEquals(10, $count);
+        self::assertEquals(10, $count);
     }
 
-    public function testDeleteQueue()
+    public function testDeleteQueue(): void
     {
         $deleted = $this->provider->deleteQueue("RiverlineWorkerBundleTest_create_v2");
 
-        $this->assertTrue($deleted);
+        self::assertTrue($deleted);
     }
 
-    public function testGetQueueOptions()
+    public function testGetQueueOptions(): void
     {
         $this->provider->createQueue("RiverlineWorkerBundleTest_queue1", ["VisibilityTimeout" => 60]);
         $queueOptions = $this->provider->getQueueOptions("RiverlineWorkerBundleTest_queue1");
 
-        $this->assertTrue(is_array($queueOptions));
-        $this->assertArrayHasKey('VisibilityTimeout', $queueOptions);
-        $this->assertEquals(60, $queueOptions['VisibilityTimeout']);
+        self::assertTrue(is_array($queueOptions));
+        self::assertArrayHasKey('VisibilityTimeout', $queueOptions);
+        self::assertEquals(60, $queueOptions['VisibilityTimeout']);
     }
 
-    public function testListQueues()
+    public function testListQueues(): void
     {
         $queues = $this->provider->listQueues("RiverlineWorkerBundleTest");
 
-        $this->assertEquals(1, count($queues));
+        self::assertCount(1, $queues);
     }
 
-    public function testQueueExists()
+    public function testQueueExists(): void
     {
         $queueExists = $this->provider->queueExists("RiverlineWorkerBundleTest_queue1");
-        $this->assertTrue($queueExists);
+        self::assertTrue($queueExists);
 
         $queueNotExists = $this->provider->queueExists("RiverlineWorkerBundleTest_queueX");
-        $this->assertFalse($queueNotExists);
+        self::assertFalse($queueNotExists);
     }
 
-    public function testUpdateQueue()
+    public function testUpdateQueue(): void
     {
         $this->provider->createQueue("RiverlineWorkerBundleTest_queue2", ["VisibilityTimeout" => 60]);
-        $queueUpdated = $this->provider->updateQueue("RiverlineWorkerBundleTest_queue2", array('ReceiveMessageWaitTimeSeconds' => '20'));
-        $this->assertTrue($queueUpdated);
+        $queueUpdated = $this->provider->updateQueue(
+            "RiverlineWorkerBundleTest_queue2",
+            ['ReceiveMessageWaitTimeSeconds' => '20']
+        );
+        self::assertTrue($queueUpdated);
 
-        $queueUpdated = $this->provider->updateQueue("RiverlineWorkerBundleTest_queue2", array('ReceiveMessageWaitTimeSeconds' => '0'));
-        $this->assertTrue($queueUpdated);
+        $queueUpdated = $this->provider->updateQueue(
+            "RiverlineWorkerBundleTest_queue2",
+            ['ReceiveMessageWaitTimeSeconds' => '0']
+        );
+        self::assertTrue($queueUpdated);
     }
 
-    public function testMessagesPerRequest()
+    public function testMessagesPerRequest(): void
     {
-        $provider = new AwsSQSv2(array(
-            'key'      => "key",
-            'secret'   => "secret",
-            'region'   => "us-east-1",
-            'version'  => "latest",
-            'endpoint' => 'http://aws:4576',
-        ), 5);
+        $provider = new AwsSQSv2(
+            [
+                'key' => "key",
+                'secret' => "secret",
+                'region' => "us-east-1",
+                'version' => "latest",
+                'endpoint' => 'http://aws:4566',
+            ], 5
+        );
 
-        $queue = $provider->createQueue("RiverlineWorkerBundleTest_create_multi", array('VisibilityTimeout' => '15'));
+        $queue = $provider->createQueue("RiverlineWorkerBundleTest_create_multi", ['VisibilityTimeout' => '15']);
 
         $queue->multiPut(['a', 'b', 'c', 'd']);
 
-        $this->assertNotNull($queue->get());
-        $this->assertNotNull($queue->get());
-        $this->assertNotNull($queue->get());
+        self::assertNotNull($queue->get());
+        self::assertNotNull($queue->get());
+        self::assertNotNull($queue->get());
 
-        $this->assertSame(0, $queue->count());
+        self::assertSame(0, $queue->count());
 
         // Destruction
         unset($provider, $queue);
 
-        $this->assertSame(1, $this->provider->count('RiverlineWorkerBundleTest_create_multi'));
+        self::assertSame(1, $this->provider->count('RiverlineWorkerBundleTest_create_multi'));
     }
 }
